Number of segments as uint32_t

+++ include/valik/shared.hpp
@@ -59,8 +59,8 @@ struct split_arguments
     std::filesystem::path meta_out{"metadata.txt"};
 
     size_t overlap{150};
-    size_t seg_count{64};
-    size_t seg_count_in{64};
+    uint32_t seg_count{64};
+    uint32_t seg_count_in{64};
     bool split_index{false};
     bool write_ref{false};
     bool write_query{false};
